Commands + More

+++ db/IFeedable.cs
@@ -1,4 +1,4 @@
 ﻿public interface IFeedable
 {
-    ushort FeedPower { get; set; }
+     int FeedPower { get; set; }
 }
